requirements/requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
Django==3.2.8
django-environ==0.8.1
django-filter==21.1
django-forms-bootstrap==3.1.0
djangorestframework==3.13.1
gunicorn==20.1.0
importlib-metadata==4.11.3
Markdown==3.3.7
psycopg2==2.9.1
pytz==2021.3
soupsieve==2.2.1
sqlparse==0.4.2
whitenoise==5.3.0
zipp==3.8.0
